+++ web/index.dart
@@ -1,34 +1,10 @@
 import 'dart:html';
+import 'compute.dart';
 
-double getValueOnInput(String idname) {
-  var input = document.querySelector(idname) as InputElement;
-  double value = 0.0;
-  try {
-    value = double.parse(input.value);
-  } catch (exception) {
-    window.alert("不正な値が入力されています：" + idname);
-  }
-  return value;
-}
-
-void writeConsole(String text) {
-  var message = new DivElement()..text = text;
-  document.body.append(message);
-}
-
-void calculate() {
-  double len0 = getValueOnInput('#length');
-  double diameter = getValueOnInput('#diameter');
-  double freq = getValueOnInput('#frequency');
-  double density = getValueOnInput('#density');
-  double young = getValueOnInput('#youngs-modulus');
-
-  writeConsole("Computed results.");
-}
 
 void main() {
   var button = document.querySelector("#calculate-btn");
-  button..onClick.listen((_) => calculate());
+  button..onClick.listen((_) => computeString());
 
   writeConsole("Loaded dart scripts.");
 }
